Move CoreError to DTO package
