+++ src/main/java/ru/team/todo/dto/linkedtasks/LinkTaskResponse.java
@@ -3,7 +3,7 @@ package ru.team.todo.dto.linkedtasks;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
-import ru.team.todo.validation.CoreError;
+import ru.team.todo.dto.CoreError;
 
 import java.util.List;
