+++ src/main/java/ru/team/todo/services/TaskService.java
@@ -7,7 +7,7 @@ import ru.team.todo.domain.User;
 import ru.team.todo.dto.tasks.*;
 import ru.team.todo.repository.TaskRepository;
 import ru.team.todo.repository.UserRepository;
-import ru.team.todo.validation.CoreError;
+import ru.team.todo.dto.CoreError;
 
 import java.util.ArrayList;
 import java.util.Collection;
